// sanity/sanity.types.ts
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol
// ...
export type USER_JOURNAL_ENTRIES_QUERYResult = {
  _id: string
  title: string | null
  content:
    | (
        | {
            children?: {
              marks?: string[]
              text?: string
              _type: 'span'
              _key: string
            }[]
            style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'normal'
            listItem?: 'bullet' | 'number'
            markDefs?: {
              href?: string
              _type: 'link'
              _key: string
            }[]
            level?: number
            _type: 'block'
            _key: string
          }
        | {
            asset?: {
              _ref: string
              _type: 'reference'
              _weak?: boolean
              [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
            }
            media?: unknown
            hotspot?: SanityImageHotspot
            crop?: SanityImageCrop
            alt?: string
            caption?: string
            _type: 'image'
            _key: string
          }
      )[]
    | null
  mood: 'happy' | 'neutral' | 'sad' | 'very-happy' | 'very-sad' | null
  createdAt: string | null
  aiGeneratedCategory: {
    title: string | null
    color: string | null
  } | null
}[]